requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
boto3==1.28.85
Pillow==10.0.1
requests==2.31.0
python-dateutil==2.8.2
